code/pyproject.toml
[tool.poetry]
name = "Swompy"
version = "0.1.0"
description = ""
authors = ["Your Name you@example.com"]
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
pyyaml = {version = "^6.0.3", python = "3.10"}
schema = "^0.7.8"
gitpython = "^3.1.45"
pydantic-settings = "^2.12.0"
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
